refactor: move parsing lists of media querry to submodule
